MVR-204 CSS styles broken into individual *.module.css files on a per component basis.
