+++ src/components/footer/Footer.module.css
@@ -7,3 +7,10 @@ footer a:hover,
 footer a.active {
     text-decoration: underline;
 }
+
+/* Phone */
+@media screen and (max-width: 480px) {
+    footer {
+        font-size: 0.85rem;
+    }
+}
