+++ src/components/header/Header.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import {NavLink, useSubmit} from "react-router-dom";
-import classes from '../main.module.css';
-import {getSubject, isAdminUser} from "../utils/auth";
+import mainClasses from '../../main.module.css';
+import classes from './Header.module.css';
+import {getSubject, isAdminUser} from "../../utils/auth";
 
 const Header = () => {
     const isAdmin = isAdminUser();
@@ -14,8 +15,8 @@ const Header = () => {
 
     return (
         <>
-            <header className={classes.header}></header>
-            <nav className={classes.navigation}>
+            <header className={mainClasses.header}></header>
+            <nav className={mainClasses.navigation}>
                 <ul className={classes.nav_list}>
                     <li><NavLink to="/" className={({isActive}) =>
                         isActive ? classes.active : undefined
