+++ src/components/recipe/Ingredients.js
@@ -2,7 +2,7 @@ import React from 'react';
 import classes from "../../main.module.css";
 import IngredientInput from "../admin/IngredientInput";
 import {isAdminUser} from "../../utils/auth";
-import DraggableList from "./DraggableList";
+import DraggableList from "./draggablelist/DraggableList";
 
 const Ingredients = (props) => {
     const isAdmin = isAdminUser();
