+++ src/components/recipe/draggablelist/DraggableList.module.css
@@ -1,0 +1,5 @@
+.draggable-list-item {
+    font-size: 1.125rem;
+    line-height: 1.5rem;
+    padding: 0.25rem;
+}
