+++ src/components/recipe/recipeslist/RecipeList.module.css
@@ -1,0 +1,7 @@
+.recipes-list {
+    padding: 0.5rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+}
