+++ src/components/recipe/recipeslist/RecipesList.js
@@ -1,14 +1,15 @@
 import React from 'react';
-import classes from "../../main.module.css";
+import classes from './RecipeList.module.css';
+import mainClasses from '../../../main.module.css';
 import {Link} from "react-router-dom";
 import _ from 'lodash';
 
-const RecipesList = (props) => {
+const RecipesList = props => {
     return (
-        <section className={classes['recipe-search-results']}>
+        <section className={classes['recipes-list']}>
             <ul>
                 {!_.isEmpty(props.recipes) && props.recipes.map(recipe =>
-                    <li key={recipe.id} className={classes.description}>
+                    <li key={recipe.id} className={mainClasses.description}>
                         <Link to={`/recipes/${recipe.id}`}>
                             {recipe.name}
                         </Link>&nbsp;
