+++ src/pages/RootLayout.js
@@ -1,5 +1,5 @@
 import React, {useEffect} from 'react';
-import Header from "../components/Header";
+import Header from "../components/header/Header";
 import {Outlet, useLoaderData, useSubmit} from "react-router-dom";
 import classes from '../main.module.css';
 import {ToastContainer} from "react-toastify";
