+++ src/pages/search/Search.module.css
@@ -1,0 +1,5 @@
+.tagList-search {
+    padding: 0.5rem;
+    display: flex;
+    justify-content: center;
+}
